Ajout des filtres

+++ SR03-JEE/src/projectSR03/servlets/ManageQuestServlet.java
@@ -12,6 +12,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import projectSR03.DAO.QuestionDAO;
 import projectSR03.DAO.QuestionnaireDAO;
+import projectSR03.DAO.UserDAO;
 import projectSR03.beans.QuestionBean;
 import projectSR03.beans.QuestionnaireBean;
 
@@ -36,31 +37,41 @@ public class ManageQuestServlet extends HttpServlet {
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		String id = req.getParameter("q_id");
+		String filter = req.getParameter("filter");
 		ArrayList<QuestionBean> q = null;
 				
 		// Delete des réponses car non lié par clé primaire directe
 		// Nécessite de trouver le questionnaire qu'on veut delete
-		for(QuestionnaireBean b : questionnaires) {
-			if(String.valueOf(b.getId()).equals(id)) {
-				q = b.getQuestions();
-				questionnaires.remove(b);
-				break;
+		if(id != null) {
+			for(QuestionnaireBean b : questionnaires) {
+				if(String.valueOf(b.getId()).equals(id)) {
+					q = b.getQuestions();
+					questionnaires.remove(b);
+					break;
+				}
+			}
+			if(q == null) { // C'est possiblement la merde
+				System.out.println("Problème");
+			} else {
+				for(QuestionBean question : q)
+				{
+					QuestionDAO.deleteAnswers(question);
+					QuestionDAO.deleteQuestion(question);
+				}
+				
+				// Puis delete du questionnaire 
+				QuestionnaireDAO.deleteQuestionnaire(id);
 			}
 		}
-		if(q == null) { // C'est possiblement la merde
-			System.out.println("Problème");
-		} else {
-			for(QuestionBean question : q)
-			{
-				QuestionDAO.deleteAnswers(question);
-				QuestionDAO.deleteQuestion(question);
-			}
-			
-			// Puis delete du questionnaire 
-			QuestionnaireDAO.deleteQuestionnaire(id);
+		if(filter != null) {
+			filterData(filter);
 		}
 		// Renvoi des données pour reload de la page ? 
 		req.setAttribute("listQuest", questionnaires);
 		this.getServletContext().getRequestDispatcher( QUEST_MANAGE ).forward( req, resp );
+	}
+	
+	public void filterData(String filter) {
+		questionnaires = QuestionnaireDAO.getQuestionnaireByName(filter);
 	}
 }
